// gesfut-front/src/app/core/models/teamRequest.ts
// Generated by https://quicktype.io

export interface TeamRequest {
    name:    string;
    color:   string;
    players: PlayerRequest[];
}

export interface PlayerRequest {
    name:         string;
    lastName:     string;
    isCaptain:    boolean;
    isGoalKeeper: boolean;
    number:       number;
}